@@ -9,8 +9,8 @@
     "prebuild": "node ./scripts/prebuild.js",
     "build": "cross-env NODE_ENV=production next build",
     "start": "next start",
-    "lint": "npx eslint . --format=stylish && next lint",
-    "lint:fix": "npx eslint . --fix && next lint --fix",
+    "lint": "npx eslint . --format=stylish",
+    "lint:fix": "npx eslint . --fix",
     "type-check": "tsc --noEmit",
     "format": "npx prettier --write .",
     "format:check": "npx prettier --check .",
@@ -19,7 +19,6 @@
     "spell:fix": "cspell \"**/*.{ts,tsx,js,jsx,md,json}\" --show-suggestions",
     "clean:all": "rimraf .next out .open-next .wrangler",
     "analyze": "ANALYZE=true npm run build",
-    "analyze:bundle": "CODECOV_TOKEN=${CODECOV_TOKEN} npm run build",
     "export": "next export",
     "test": "vitest run",
     "test:watch": "vitest",
@@ -30,7 +29,7 @@
     "prepreview": "rimraf .open-next .wrangler",
     "preview": "opennextjs-cloudflare build && opennextjs-cloudflare preview",
     "deploy": "opennextjs-cloudflare build && opennextjs-cloudflare deploy",
-    "cf-typegen": "wrangler types --env-interface CloudflareEnv cloudflare-env.d.ts",
+    "cf-typegen": "wrangler types --env-interface CloudflareEnv ./cloudflare-env.d.ts",
     "build:sitedata": "npm run build:blog && npm run build:page-dates",
     "build:blog": "node scripts/build-blog-data.js",
     "build:page-dates": "node scripts/build-page-dates.js",
@@ -42,7 +41,7 @@
     "dompurify": "^3.2.7",
     "framer-motion": "^12.23.22",
     "lucide-react": "^0.544.0",
-    "next": "15.4.7",
+    "next": "15.5.4",
     "next-intl": "^4.3.9",
     "node-html-parser": "^7.0.1",
     "prismjs": "^1.30.0",
@@ -56,7 +55,6 @@
     "unist-util-visit": "^5.0.0"
   },
   "devDependencies": {
-    "@codecov/nextjs-webpack-plugin": "^1.9.1",
     "@cspell/dict-sv": "^2.3.2",
     "@eslint/eslintrc": "^3.3.1",
     "@eslint/js": "^9.37.0",
